// src/app/types/product.ts
export type Product = {
  id: number;
  name: string;
  slug: string;
  permalink: string;
  date_created: string;
  date_created_gmt: string;
  date_modified: string;
  date_modified_gmt: string;
  type: string;
  status: string;
  featured: boolean;
  catalog_visibility: string;
  description: string;
  short_description: string;
  sku: string;
  price: string;
  regular_price: string;
  sale_price: string;
  date_on_sale_from: string | null;
  date_on_sale_from_gmt: string | null;
  date_on_sale_to: string | null;
  date_on_sale_to_gmt: string | null;
  on_sale: boolean;
  purchasable: boolean;
  total_sales: number;
  virtual: boolean;
  downloadable: boolean;
  downloads: Array<any>;
  download_limit: number;
  download_expiry: number;
  external_url: string;
  button_text: string;
  tax_status: string;
  tax_class: string;
  manage_stock: boolean;
  stock_quantity: number | null;
  backorders: string;
  backorders_allowed: boolean;
  backordered: boolean;
  low_stock_amount: number | null;
  sold_individually: boolean;
  weight: string;
  dimensions: {
    length: string;
    width: string;
    height: string;
  };
  shipping_required: boolean;
  shipping_taxable: boolean;
  shipping_class: string;
  shipping_class_id: number;
  reviews_allowed: boolean;
  average_rating: string;
  rating_count: number;
  upsell_ids: Array<number>;
  cross_sell_ids: Array<number>;
  parent_id: number;
  purchase_note: string;
  categories: Array<any>;
  tags: Array<any>;
  images: Array<any>;
  attributes: Array<any>;
  default_attributes: Array<any>;
  variations: Array<number>;
  grouped_products: Array<any>;
  menu_order: number;
  price_html: string;
  related_ids: Array<number>;
  meta_data: Array<any>;
  stock_status: string;
  has_options: boolean;
  post_password: string;
  google_listings_and_ads__channel_visibility: {
    is_visible: boolean;
    channel_visibility: string;
    sync_status: string;
    issues: Array<any>;
  };
  _links: {
    self: Array<{ href: string }>;
    collection: Array<{ href: string }>;
  };
};

export type Products = Product[];